express server setup. starting dockerization.
