+++ .vscode/launch.json
@@ -11,7 +11,7 @@
             "skipFiles": [
                 "<node_internals>/**"
             ],
-            "program": "${workspaceFolder}/index.js",
+            "program": "${workspaceFolder}/server/index.js",
             "outFiles": [
                 "${workspaceFolder}/**/*.js"
             ]
