+++ client/src/components/Omegaphone/Hat.ts
