+++ client/src/components/Omegaphone/Kick.ts
