+++ client/src/components/Omegaphone/synths/index.ts
