+++ client/src/components/Omegaphone/synths/kick-synth.ts
