+++ client/src/utils/rand.ts
