--- src/components/Iota.ts
@@ -1,64 +1,0 @@
-import * as THREE from "three";
-import GlowMesh from "./GlowMesh";
-import { iotaPalette } from "../colors";
-import { IotaPrefs, IotaSystemOptions } from "../types";
-import getInitialPrefs from "./prefs";
-import { rand } from "../utils";
-
-export default class Iota extends THREE.Object3D {
-    geometry: THREE.SphereBufferGeometry;
-    material: THREE.MeshLambertMaterial;
-    mesh: THREE.Mesh;
-    prefs: IotaPrefs;
-    velocity: number[];
-
-    constructor({ worldSize }: IotaSystemOptions) {
-        super();
-
-        this.geometry = new THREE.SphereBufferGeometry(5, 32, 32);
-        this.material = new THREE.MeshLambertMaterial({
-            color: "#277ec9",
-            transparent: true,
-            opacity: 1.0,
-        });
-
-        this.mesh = new THREE.Mesh(this.geometry, this.material);
-
-        const glowMesh = new GlowMesh(this.mesh);
-
-        this.mesh.add(glowMesh.object3d);
-
-        const [x, y, z] = this.getStartPosition(worldSize);
-        const color = this.getRandomColor();
-
-        this.mesh.position.set(x, y, z);
-        this.velocity = this.getVelocity();
-        this.material.color.set(color);
-        this.material.reflectivity = 0.2;
-        this.prefs = getInitialPrefs();
-    }
-
-    getStartPosition(worldSize: number): number[] {
-        const x = rand(-worldSize, worldSize);
-        const y = rand(-worldSize, worldSize);
-        const z = rand(-worldSize, worldSize);
-
-        return [x, y, z];
-    }
-
-    getPaletteColor(): number {
-        return iotaPalette[Math.floor(rand(0, iotaPalette.length))];
-    }
-
-    getRandomColor(): string {
-        const h = Math.ceil(rand(0, 360));
-        const s = Math.ceil(rand(20, 100));
-        const l = Math.ceil(rand(20, 80));
-
-        return `hsl(${h}, ${s}%, ${l}%)`;
-    }
-
-    getVelocity(): number[] {
-        return [rand(-1, 1), rand(-1, 1), rand(-1, 1)];
-    }
-}
